Nouvelle table

+++ Data_synthese/Data_synthese/Classes/DatabaseObject.cs
@@ -3,20 +3,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Entites_Synthese;
-
+using System.Data.Entity.Validation;
 
 namespace Data_synthese.Classes
 {
     public class DatabaseObject : IDisposable
     {
+        private Session _session = Session.getInstance();
         public static int erreur;
+
         public DatabaseObject()
         {
             this.dbContext = new synthese_dbEntities();
             Crypteur = new Encription();
         }
-
-
 
         #region Propriétés
 
@@ -42,10 +42,65 @@ namespace Data_synthese.Classes
 
         public void SaveChanges()
         {
-            this.dbContext.SaveChanges();
+            SaveChanges(this.dbContext);
+        }
+        /// <summary>
+        /// Wrapper for SaveChanges adding the Validation Messages to the generated exception
+        /// </summary>
+        /// <param name="context">The context.</param>
+        private void SaveChanges(synthese_dbEntities context)
+        {
+            try
+            {
+                context.SaveChanges();
+            }
+            catch (DbEntityValidationException ex)
+            {
+                StringBuilder sb = new StringBuilder();
+
+                foreach (var failure in ex.EntityValidationErrors)
+                {
+                    sb.AppendFormat("{0} failed validation\n", failure.Entry.Entity.GetType());
+                    foreach (var error in failure.ValidationErrors)
+                    {
+                        sb.AppendFormat("- {0} : {1}", error.PropertyName, error.ErrorMessage);
+                        sb.AppendLine();
+                    }
+                }
+
+                throw new DbEntityValidationException(
+                    "Entity Validation Failed - errors follow:\n" + sb, ex); // Add the original exception as the innerException
+            }
+        }
+        #endregion
+
+        public Usager_Entite Login(string pUser, string pPassword)
+        {
+
+            string hash = string.Empty;
+            Usager_Entite usager = GetUsager(pUser);
+            if (usager != null)
+            {
+                hash = Encription.GetHash(pPassword);
+
+                if (Encription.VerifiyHash(pPassword, usager.MotDePasse))
+                {
+                    _session.usager = usager;
+                    return usager;
+                }
+            }
+            else
+
+                throw new Exception(Constantes.ERREUR_LOGIN);
+
+            return null;
         }
 
-        #endregion
+        public void Logout()
+        {
+
+            _session.usager = null;
+        }
 
         #region Usagers
 
@@ -86,12 +141,11 @@ namespace Data_synthese.Classes
                         }
             }
 
-            //TODO: il manque des propriétés dans l'usager...
             if (userRow != null)
             {
                 usag = new Usager_Entite();
                 usag.ID = userRow.Id;
-
+                usag.EstAdministrateur = userRow.Administrateur;
                 usag.Nom = userRow.Nom;
                 usag.MotDePasse = userRow.MotPasse;
                 usag.NomUsager = userRow.NomUsager;
@@ -136,76 +190,84 @@ namespace Data_synthese.Classes
                                           bool pEstAdministrateur,
                                           String pCourriel)
         {
-            Usager_Entite usag;
-            Encription crypteur = new Encription();
+            Usager_Entite usag = null;
 
-            // On s'assure que le username n'existe pas déjà
-            var userNameRow = (from row in dbContext.usager.Local
-                               where row.NomUsager == pUsername
-                               select row);
-            if ((userNameRow != null) && (userNameRow.ToList().Count == 0))
-                userNameRow = (from row in dbContext.usager
-                               where row.NomUsager == pUsername
-                               select row);
-            if ((userNameRow != null) && (userNameRow.ToList().Count > 0))
+            if ((_session.usager != null) && (_session.usager.EstAdministrateur == true))
             {
-                // err //définir des erreurs et finaliser le test
-                throw new Exceptions.UsagerExistantException();
+
+                Encription crypteur = new Encription();
+
+                // On s'assure que le username n'existe pas déjà
+                var userNameRow = (from row in dbContext.usager.Local
+                                   where row.NomUsager == pUsername
+                                   select row);
+                if ((userNameRow != null) && (userNameRow.ToList().Count == 0))
+                    userNameRow = (from row in dbContext.usager
+                                   where row.NomUsager == pUsername
+                                   select row);
+                if ((userNameRow != null) && (userNameRow.ToList().Count > 0))
+                {
+                    // err //définir des erreurs et finaliser le test
+                    throw new Exception(Classes.Constantes.ERREUR_NOM_UTILISATEUR_EXISTANT);
+
+                }
+                var courrielRow = (from row in dbContext.usager.Local
+                                   where row.Courriel == pCourriel
+                                   select row);
+                if ((courrielRow != null) && (courrielRow.ToList().Count == 0))
+                    courrielRow = (from row in dbContext.usager
+                                   where row.Courriel == pCourriel
+                                   select row);
+
+                if ((courrielRow != null) && (courrielRow.ToList().Count > 0))
+                    throw new Exception(Constantes.ERREUR_COURRIEL_EXISTANT);
+
+                // On s'assure que le courriel n'existe pas déjà
+
+                // On va chercher le dernier ID
+                var userRow = (from row in dbContext.usager.Local
+                               orderby row.Id descending
+                               select row).ToList();
+                if (userRow.Count == 0)
+                    userRow = (from row in dbContext.usager
+                               orderby row.Id descending
+                               select row).ToList();
+                int id;
+                if (userRow.Count > 0)
+                    id = userRow.First().Id;
+                else
+                    id = 0;
+
+                string passwordHash = Encription.GetHash(pPassword);
+
+                usager usagerDB = new usager()
+                {
+                    Nom = pNom,
+                    NomUsager = pUsername,
+                    MotPasse = passwordHash,
+                    Administrateur = pEstAdministrateur,
+                    Courriel = pCourriel,
+                    Id = (id += 1)
+                };
+                dbContext.usager.Add(usagerDB);
+
+                usag = new Usager_Entite(usagerDB.Id,
+                    pNom,
+                    pUsername,
+                    pPassword,
+                    pEstAdministrateur,
+                    pCourriel);
 
             }
-            var courrielRow = (from row in dbContext.usager.Local
-                               where row.Courriel == pCourriel
-                               select row);
-            if ((userNameRow != null) && (userNameRow.ToList().Count == 0))
-                courrielRow = (from row in dbContext.usager
-                               where row.Courriel == pCourriel
-                               select row);
-
-            if ((userNameRow != null) && (userNameRow.ToList().Count > 0))
-                throw new Exceptions.CourrielExistantException();
-
-            // On s'assure que le courriel n'existe pas déjà
-
-            // On va chercher le dernier ID
-            var userRow = (from row in dbContext.usager.Local
-                           orderby row.Id descending
-                           select row).ToList();
-            if (userRow.Count == 0)
-                userRow = (from row in dbContext.usager
-                           orderby row.Id descending
-                           select row).ToList();
-            int id;
-            if (userRow.Count > 0)
-                id = userRow.First().Id;
             else
-                id = 0;
-
-            string passwordHash = Encription.GetHash(pPassword);
-
-            usager usagerDB = new usager()
-            {
-                Nom = pNom,
-                NomUsager = pUsername,
-                MotPasse = passwordHash,
-                Administrateur = pEstAdministrateur,
-                Courriel = pCourriel,
-                Id = (id += 1)
-            };
-            dbContext.usager.Add(usagerDB);
-
-            usag = new Usager_Entite(usagerDB.Id,
-                pNom,
-                pUsername,
-                pPassword,
-                pEstAdministrateur,
-                pCourriel);
+                throw new Exception(Constantes.ERREUR_ADMIN_REQUIS);
 
             return usag;
         }
 
         #endregion
 
-
+        #region " GetUsager "
         /// <summary>
         /// Retourne l'usager ayant l'ID= pID ou un usager vide s'il n'est pas dans la BD
         /// </summary>
@@ -235,46 +297,87 @@ namespace Data_synthese.Classes
 
             return usag;
         }
+        #endregion
 
-
-
+        #region " DeleteUsager "
         public Boolean DeleteUsager(int pUsagerID)
         {
-            var UsagerRow = (from row in dbContext.usager.Local
-                             where row.Id == pUsagerID
-                             select row).FirstOrDefault();
-            if (UsagerRow == null)
-                UsagerRow = (from row in dbContext.usager
-                             where row.Id == pUsagerID
-                             select row).FirstOrDefault();
-            if (UsagerRow != null)
+            if ((_session.usager != null) && (_session.usager.EstAdministrateur == true))
             {
-                dbContext.usager.Remove(UsagerRow);
+                var UsagerRow = (from row in dbContext.usager.Local
+                                 where row.Id == pUsagerID
+                                 select row).FirstOrDefault();
+                if (UsagerRow == null)
+                    UsagerRow = (from row in dbContext.usager
+                                 where row.Id == pUsagerID
+                                 select row).FirstOrDefault();
+                if (UsagerRow != null)
+                {
+                    dbContext.usager.Remove(UsagerRow);
 
-                return true;
+                    return true;
+                }
             }
+            else
+                throw new Exception(Constantes.ERREUR_ADMIN_REQUIS);
             return false;
         }
 
 
         public Boolean DeleteUsager(string pUserName)
         {
-            var UsagerRow = (from row in dbContext.usager.Local
-                             where row.NomUsager == pUserName
-                             select row).FirstOrDefault();
-            if (UsagerRow == null)
-                UsagerRow = (from row in dbContext.usager
-                             where row.NomUsager == pUserName
-                             select row).FirstOrDefault();
-            if (UsagerRow != null)
+            if ((_session.usager != null) && (_session.usager.EstAdministrateur == true))
             {
-                dbContext.usager.Remove(UsagerRow);
+                var UsagerRow = (from row in dbContext.usager.Local
+                                 where row.NomUsager == pUserName
+                                 select row).FirstOrDefault();
+                if (UsagerRow == null)
+                    UsagerRow = (from row in dbContext.usager
+                                 where row.NomUsager == pUserName
+                                 select row).FirstOrDefault();
+                if (UsagerRow != null)
+                {
+                    dbContext.usager.Remove(UsagerRow);
 
-                return true;
+                    return true;
+                }
             }
+            else
+                throw new Exception(Constantes.ERREUR_ADMIN_REQUIS);
             return false;
         }
+        #endregion
 
+
+        #region "UpdateUsager"
+
+        public Usager_Entite UpdateUsager(Usager_Entite pUsager)
+        {
+
+            Encription crypteur = new Encription();
+            var usagerRow = (from row in dbContext.usager.Local
+                             where row.Id == pUsager.ID
+                             select row).FirstOrDefault();
+            if (usagerRow == null)
+                usagerRow = (from row in dbContext.usager
+                             where row.Id == pUsager.ID
+                             select row).FirstOrDefault();
+
+            if (usagerRow != null)
+            {
+                usagerRow.Nom = pUsager.Nom;
+                usagerRow.NomUsager = pUsager.NomUsager;
+                usagerRow.MotPasse = pUsager.MotDePasse;
+                usagerRow.Administrateur = pUsager.EstAdministrateur;
+                usagerRow.Courriel = pUsager.Courriel;
+            }
+            else
+                pUsager.MessageErreur = Constantes.ERREUR_USAGER_INEXISTANT;
+
+            return pUsager;
+
+        }
+        #endregion
 
         #endregion
